# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Trançify SaaS multi-tenant scheduling API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /availability:
    get:
      operationId: getMyAvailability
      tags: [availability]
      summary: Get availability config for current tenant
      responses:
        "200":
          description: Availability config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Availability"
    put:
      operationId: updateAvailability
      tags: [availability]
      summary: Update availability config
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAvailabilityRequest"
      responses:
        "200":
          description: Updated availability
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Availability"
  /availability/public/{tenantId}:
    get:
      operationId: getPublicAvailability
      tags: [availability]
      summary: Get available time slots for a tenant on a given date
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: serviceId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Available slots
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailableSlots"
  /appointments:
    get:
      operationId: getMyAppointments
      tags: [appointments]
      summary: Get appointments for current tenant
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, cancelled, completed]
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
  /appointments/{id}:
    get:
      operationId: getAppointment
      tags: [appointments]
      summary: Get a single appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    patch:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update an appointment (tenant)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAppointmentRequest"
      responses:
        "200":
          description: Updated appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    delete:
      operationId: deleteAppointment
      tags: [appointments]
      summary: Delete an appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /appointments/book:
    post:
      operationId: bookAppointment
      tags: [appointments]
      summary: Book a new appointment (client, public)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookAppointmentRequest"
      responses:
        "201":
          description: Appointment booked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "400":
          description: Slot unavailable or validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /appointments/{id}/cost:
    patch:
      operationId: updateAppointmentCost
      tags: [appointments]
      summary: Update material cost for an appointment (tenant)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCostRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    Availability:
      type: object
      properties:
        id:
          type: string
        tenantId:
          type: string
        availableDays:
          type: array
          items:
            type: integer
            description: Day of week 0=Sun, 1=Mon, ..., 6=Sat
        startTime:
          type: string
          description: HH:MM format
        endTime:
          type: string
          description: HH:MM format
        slotIntervalMinutes:
          type: integer
        breakAfterMinutes:
          type: integer
          description: Rest time after each appointment in minutes
        maxAppointmentsPerDay:
          type: integer
        blockedDates:
          type: array
          items:
            type: string
            format: date
      required: [id, tenantId, availableDays, startTime, endTime, slotIntervalMinutes, breakAfterMinutes, maxAppointmentsPerDay]
    UpdateAvailabilityRequest:
      type: object
      properties:
        availableDays:
          type: array
          items:
            type: integer
        startTime:
          type: string
        endTime:
          type: string
        slotIntervalMinutes:
          type: integer
        breakAfterMinutes:
          type: integer
        maxAppointmentsPerDay:
          type: integer
        blockedDates:
          type: array
          items:
            type: string
            format: date
    AvailableSlots:
      type: object
      properties:
        date:
          type: string
          format: date
        slots:
          type: array
          items:
            type: string
            description: Available time slots in HH:MM format
        available:
          type: boolean
      required: [date, slots, available]
    Appointment:
      type: object
      properties:
        id:
          type: string
        tenantId:
          type: string
        serviceId:
          type: string
        serviceName:
          type: string
        clientName:
          type: string
        clientAge:
          type: integer
          nullable: true
        clientPhone:
          type: string
          nullable: true
        hairDescription:
          type: string
          nullable: true
        referencePhotos:
          type: array
          items:
            type: string
        paymentMethod:
          type: string
          enum: [pix, card, cash]
        braidSize:
          type: string
          enum: [mid_back, waist_butt]
        servicePrice:
          type: number
        materialCost:
          type: number
          nullable: true
        profit:
          type: number
          nullable: true
        date:
          type: string
          format: date
        time:
          type: string
          description: HH:MM
        status:
          type: string
          enum: [pending, confirmed, cancelled, completed]
        token:
          type: string
          description: Client editing token
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, tenantId, serviceId, serviceName, clientName, paymentMethod, braidSize, servicePrice, date, time, status, token, createdAt]
    BookAppointmentRequest:
      type: object
      properties:
        tenantId:
          type: string
        serviceId:
          type: string
        clientName:
          type: string
        clientAge:
          type: integer
        clientPhone:
          type: string
        hairDescription:
          type: string
        referencePhotos:
          type: array
          items:
            type: string
        paymentMethod:
          type: string
          enum: [pix, card, cash]
        braidSize:
          type: string
          enum: [mid_back, waist_butt]
        date:
          type: string
          format: date
        time:
          type: string
        notes:
          type: string
      required: [tenantId, serviceId, clientName, paymentMethod, braidSize, date, time]
    UpdateAppointmentRequest:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, cancelled, completed]
        date:
          type: string
          format: date
        time:
          type: string
        notes:
          type: string
    UpdateCostRequest:
      type: object
      properties:
        materialCost:
          type: number
      required: [materialCost]
...